refactor(repository): domain.repository 패키지 추가
