+++ src/main/java/com/basketclub/domain/repository/PlayerRepository.java
@@ -1,5 +1,6 @@
-package com.basketclub.domain;
+package com.basketclub.domain.repository;
 
+import com.basketclub.domain.Player;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
