+++ src/main/java/com/basketclub/domain/repository/ReviewRepository.java
@@ -1,5 +1,6 @@
-package com.basketclub.domain;
+package com.basketclub.domain.repository;
 
+import com.basketclub.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
