+++ src/test/java/com/basketclub/domain/PlayerRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.basketclub.domain;
 
+import com.basketclub.domain.repository.PlayerRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
